feat: replace google analytics with vercel analytics
- install @vercel/analytics package
- create new vercelanalytics component
- remove old google analytics components
- update layout to use vercel analytics
- update web-vitals components to use vercel analytics
- fix syntax error in hero-section component
- improve performance and privacy compliance
