+++ package.json
@@ -35,6 +35,7 @@
     "@sanity/vision": "^4.2.0",
     "@studio-freight/lenis": "^1.0.42",
     "@types/bcryptjs": "^2.4.6",
+    "@vercel/analytics": "^1.5.0",
     "ai": "^4.3.19",
     "bcryptjs": "^3.0.2",
     "class-variance-authority": "^0.7.1",
